refactor: extract shared OAuth credentials helper for google and microsoft
