+++ src/integrations/google/utils.ts
@@ -2,10 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { refreshAccessToken } from '@/lib/auth'
-import type { HttpClient } from '@/lib/http'
-import { getIntegrationCredentials, updateIntegrationCredentials } from '@/dal'
-import { getDb } from '@/db/database'
 import type { DraftEmailParams } from './tools'
 
 // =============================================================================
@@ -121,63 +117,6 @@ export const buildRawMessage = (params: DraftEmailParams): string => {
   return btoa(bytes.reduce((s, b) => s + String.fromCharCode(b), ''))
     .replace(/\+/g, '-')
     .replace(/\//g, '_')
-}
-
-// =============================================================================
-// AUTH UTILITY FUNCTIONS
-// =============================================================================
-
-/**
- * Retrieve stored Google OAuth credentials from settings.
- * Throws if the integration has not been connected yet or the stored value is malformed.
- */
-export const getGoogleCredentials = async (): Promise<{
-  access_token: string
-  refresh_token?: string
-  expires_at?: number
-}> => {
-  const db = getDb()
-  const row = await getIntegrationCredentials(db, 'google')
-  if (!row) {
-    throw new Error('Google integration not connected')
-  }
-  return row.credentials
-}
-
-/**
- * Ensure that we have a valid Google OAuth access token, refreshing it if necessary.
- * If the token is refreshed, the stored credentials are updated automatically.
- */
-export const ensureValidGoogleToken = async (
-  httpClient: HttpClient,
-  credentials: {
-    access_token: string
-    refresh_token?: string
-    expires_at?: number
-  },
-): Promise<string> => {
-  const now = Date.now()
-  // If the token is still valid for at least 1 minute, reuse it
-  if (credentials.expires_at && credentials.expires_at - 60_000 > now) {
-    return credentials.access_token
-  }
-
-  if (!credentials.refresh_token) {
-    throw new Error('Access token expired and no refresh token available')
-  }
-
-  const newTokens = await refreshAccessToken(httpClient, 'google', credentials.refresh_token)
-
-  const updated = {
-    ...credentials,
-    access_token: newTokens.access_token,
-    expires_at: Date.now() + newTokens.expires_in * 1000,
-  }
-
-  const db = getDb()
-  await updateIntegrationCredentials(db, 'google', updated)
-
-  return updated.access_token
 }
 
 // =============================================================================
